# apps/server/swagger.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"PingController_ping","parameters":[],"responses":{"200":{"description":""}},"tags":["Public"]}},"/register":{"post":{"operationId":"AuthenticationController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/login":{"post":{"operationId":"AuthenticationController_logIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/logout":{"post":{"operationId":"AuthenticationController_logOut","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/me":{"get":{"operationId":"AuthenticationController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Authentication"]}},"/reset-password":{"post":{"operationId":"AuthenticationController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/refresh":{"get":{"operationId":"AuthenticationController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/user":{"patch":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/email-confirmation/confirm":{"post":{"operationId":"EmailConfirmationController_confirm","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailDto"}}}},"responses":{"201":{"description":""}},"tags":["Email Confirmation"]}},"/email-confirmation/resend-confirmation-link":{"post":{"operationId":"EmailConfirmationController_resendConfirmationLink","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendEmailResponseDto"}}}}},"tags":["Email Confirmation"]}},"/google-authentication":{"post":{"operationId":"GoogleAuthenticationController_authenticate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenVerificationDto"}}}},"responses":{"201":{"description":""}},"tags":["Google Authentication"]}},"/dashboard/summary":{"get":{"operationId":"DashboardController_getSummary","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetSummaryResponse"}}}}},"tags":["Dashboard"]}},"/dashboard/users":{"get":{"operationId":"DashboardController_getUsers","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersResponse"}}}}},"tags":["Dashboard"]}}},"info":{"title":"Dashboard","description":"","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:4000"}],"components":{"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"avatar":{"type":"string"},"isEmailConfirmed":{"type":"boolean"},"isRegisteredWithGoogle":{"type":"boolean"},"lastTimeSendEmailConfirmation":{"format":"date-time","type":"string"}},"required":["id","email","name","avatar","isEmailConfirmed","isRegisteredWithGoogle","lastTimeSendEmailConfirmation"]},"UpdatePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"ConfirmEmailDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"ResendEmailResponseDto":{"type":"object","properties":{"lastTimeSendEmailConfirmation":{"format":"date-time","type":"string"},"resendTimeConfig":{"type":"number"}},"required":["lastTimeSendEmailConfirmation","resendTimeConfig"]},"TokenVerificationDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"GetSummaryResponse":{"type":"object","properties":{"totalUser":{"type":"number"},"totalActiveUser":{"type":"number"},"averageActiveUser":{"type":"number"}},"required":["totalUser","totalActiveUser","averageActiveUser"]},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"createdDate":{"format":"date-time","type":"string"},"loginCount":{"type":"number"},"lastSessionTimestamp":{"format":"date-time","type":"string"},"isRegisteredWithGoogle":{"format":"date-time","type":"string"}},"required":["id","name","email","createdDate","loginCount","lastSessionTimestamp","isRegisteredWithGoogle"]},"Pagination":{"type":"object","properties":{"total":{"type":"number"},"page":{"type":"number"},"limit":{"type":"number"}},"required":["total","page","limit"]},"GetUsersResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["data","pagination"]}}}}